refactor(tests): extract virt fixtures from tests/conftest.py
CNV-80951 — move virt fixtures into domain modules under
tests/fixtures/virt/, rebased onto main (prior conftest extractions already merged).

Modules:
- instance_types.py — instance type and VM preference param fixtures
- migration.py — migration policy, migrated_vm_multiple_times,
  is_postcopy_migration_bug_open
- osinfo.py — libosinfo DB download fixtures
- ssh.py — SSH key generation fixtures
- templates.py — base_templates
- virtual_machines.py — all VM creation fixtures and rhel_latest_os_params
- vm_actions.py — ping_process_in_rhel_os (per review feedback)

Register all seven modules in pytest_plugins in root conftest.py.
